refactor: request 삭제 및 Service 파라미터 통일
