+++ src/main/java/com/example/springlogin/member/param/LoginParam.java
@@ -1,11 +1,11 @@
-package com.example.springlogin.member.request;
+package com.example.springlogin.member.param;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class LoginRequest {
+public class LoginParam {
     private String email;
     private String password;
 }
