--- src/main/java/com/example/springlogin/member/request/JoinRequest.java
@@ -1,13 +1,0 @@
-package com.example.springlogin.member.request;
-
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class JoinRequest {
-    private String email;
-    private String password;
-
-}
