+++ src/main/java/com/example/springlogin/member/service/MemberService.java
@@ -2,10 +2,9 @@ package com.example.springlogin.member.service;
 
 import com.example.springlogin.member.domain.Member;
 import com.example.springlogin.member.param.JoinParam;
-import com.example.springlogin.member.request.JoinRequest;
-import com.example.springlogin.member.request.LoginRequest;
+import com.example.springlogin.member.param.LoginParam;
 
 public interface MemberService {
     void join(JoinParam param);
-    Member login(LoginRequest loginRequest);
+    Member login(LoginParam loginRequest);
 }
